fix: Redo income calculation
