+++ policyengine_it/variables/household/income/individual/total_individual_pre_tax_income.py
@@ -2,12 +2,17 @@ from policyengine_it.model_api import *
 
 label = "Earnings"
 
+
 class total_individual_pre_tax_income(Variable):
-    # For the time being, this is equivalent to the "gross_income" variable from the
-    # example code for head of household or to "spouse_income" for spouse
+    # In a more built-out model, there would be different categories of income
+    # that all form part of this variable. However, the existing model only contains
+    # 'gross_income' (for head of household) and 'spouse income' (for spouse).
+    # This will be used as the stand-in for that, even though it only includes one
+    # subset, 'employment_income'.
     value_type = float
     entity = Person
     label = "Total individual pre-tax income"
     unit = EUR
     documentation = "Income from gainful employment"
     definition_period = YEAR
+    adds = ["employment_income"]
